CI tests with last 4 versions. Remove 1.12 and below specific code
